Struct anymap2::raw::RawMap

pub struct RawMap<A: ?Sized + UncheckedAnyExt = dyn Any> { /* private fields */ }

The raw, underlying form of a Map.

At its essence, this is a wrapper around HashMap<TypeId, Box<Any>>, with the portions that would be memory-unsafe removed or marked unsafe. Normal people are expected to use the safe Map interface instead, but there is the occasional use for this such as iteration over the contents of an Map. However, because you will then be dealing with Any trait objects, it doesn’t tend to be so very useful. Still, if you need it, it’s here.

Implementations

impl<A: ?Sized + UncheckedAnyExt> RawMap<A>

pub fn new() -> RawMap<A>

Create an empty collection.

pub fn with_capacity(capacity: usize) -> RawMap<A>

Creates an empty collection with the given initial capacity.

pub fn capacity(&self) -> usize

Returns the number of elements the collection can hold without reallocating.

pub fn reserve(&mut self, additional: usize)

Reserves capacity for at least additional more elements to be inserted in the collection. The collection may reserve more space to avoid frequent reallocations.

Panics

Panics if the new allocation size overflows usize.

pub fn shrink_to_fit(&mut self)

Shrinks the capacity of the collection as much as possible. It will drop down as much as possible while maintaining the internal rules and possibly leaving some space in accordance with the resize policy.

pub fn len(&self) -> usize

Returns the number of items in the collection.

pub fn is_empty(&self) -> bool

Returns true if there are no items in the collection.

pub fn clear(&mut self)

Removes all items from the collection. Keeps the allocated memory for reuse.

impl<A: ?Sized + UncheckedAnyExt> RawMap<A>

pub fn iter(&self) -> Iter<'_, A>

An iterator visiting all entries in arbitrary order.

Iterator element type is &Any.

pub fn iter_mut(&mut self) -> IterMut<'_, A>

An iterator visiting all entries in arbitrary order.

Iterator element type is &mut Any.

pub fn drain(&mut self) -> Drain<'_, A>

Clears the map, returning all items as an iterator.

Iterator element type is Box<Any>.

Keeps the allocated memory for reuse.

pub fn entry(&mut self, key: TypeId) -> Entry<'_, A>

Gets the entry for the given type in the collection for in-place manipulation.

pub fn get<Q: ?Sized>(&self, k: &Q) -> Option<&A>where
TypeId: Borrow<Q>,
Q: Hash + Eq,

Returns a reference to the value corresponding to the key.

The key may be any borrowed form of the map’s key type, but Hash and Eq on the borrowed form must match those for the key type.

pub fn contains_key<Q: ?Sized>(&self, k: &Q) -> boolwhere
TypeId: Borrow<Q>,
Q: Hash + Eq,

Returns true if the map contains a value for the specified key.

The key may be any borrowed form of the map’s key type, but Hash and Eq on the borrowed form must match those for the key type.

pub fn get_mut<Q: ?Sized>(&mut self, k: &Q) -> Option<&mut A>where
TypeId: Borrow<Q>,
Q: Hash + Eq,

Returns a mutable reference to the value corresponding to the key.

The key may be any borrowed form of the map’s key type, but Hash and Eq on the borrowed form must match those for the key type.

pub unsafe fn insert(&mut self, key: TypeId, value: Box<A>) -> Option<Box<A>>

Inserts a key-value pair from the map. If the key already had a value present in the map, that value is returned. Otherwise, None is returned.

Safety

It is the caller’s responsibility to ensure that the key corresponds with the type ID of the value. If they do not, memory safety may be violated.

pub fn remove<Q: ?Sized>(&mut self, k: &Q) -> Option<Box<A>>where
TypeId: Borrow<Q>,
Q: Hash + Eq,

Removes a key from the map, returning the value at the key if the key was previously in the map.

The key may be any borrowed form of the map’s key type, but Hash and Eq on the borrowed form must match those for the key type.