Type Alias anymap2::SendSyncAnyMap

pub type SendSyncAnyMap = Map<dyn Any + Send + Sync>;

Sync version

Aliased Type

struct SendSyncAnyMap { /* private fields */ }

Implementations

impl<A: ?Sized + UncheckedAnyExt> Map<A>

pub fn new() -> Map<A>

Create an empty collection.

pub fn with_capacity(capacity: usize) -> Map<A>

Creates an empty collection with the given initial capacity.

pub fn capacity(&self) -> usize

Returns the number of elements the collection can hold without reallocating.

pub fn reserve(&mut self, additional: usize)

Reserves capacity for at least additional more elements to be inserted in the collection. The collection may reserve more space to avoid frequent reallocations.

Panics

Panics if the new allocation size overflows usize.

pub fn shrink_to_fit(&mut self)

Shrinks the capacity of the collection as much as possible. It will drop down as much as possible while maintaining the internal rules and possibly leaving some space in accordance with the resize policy.

pub fn len(&self) -> usize

Returns the number of items in the collection.

pub fn is_empty(&self) -> bool

Returns true if there are no items in the collection.

pub fn clear(&mut self)

Removes all items from the collection. Keeps the allocated memory for reuse.

impl<A: ?Sized + UncheckedAnyExt> Map<A>

pub fn get<T: IntoBox<A>>(&self) -> Option<&T>

Returns a reference to the value stored in the collection for the type T, if it exists.

pub fn get_mut<T: IntoBox<A>>(&mut self) -> Option<&mut T>

Returns a mutable reference to the value stored in the collection for the type T, if it exists.

pub fn insert<T: IntoBox<A>>(&mut self, value: T) -> Option<T>

Sets the value stored in the collection for the type T. If the collection already had a value of type T, that value is returned. Otherwise, None is returned.

pub fn remove<T: IntoBox<A>>(&mut self) -> Option<T>

Removes the T value from the collection, returning it if there was one or None if there was not.

pub fn contains<T: IntoBox<A>>(&self) -> bool

Returns true if the collection contains a value of type T.

pub fn entry<T: IntoBox<A>>(&mut self) -> Entry<'_, A, T>

Gets the entry for the given type in the collection for in-place manipulation