Crate anymap2

This crate provides the AnyMap type, a safe and convenient store for one value of each type.

Modules

any

The different types of Any for use in a map.

raw

The raw form of a Map, allowing untyped access.

Structs

Map

A collection containing zero or one values for any given type and allowing convenient, type-safe access to those values.

OccupiedEntry

A view into a single occupied location in an Map.

VacantEntry

A view into a single empty location in an Map.

Enums

Entry

A view into a single location in an Map, which may be vacant or occupied.

Type Definitions

AnyMap

The most common type of Map: just using Any.

SendSyncAnyMap

Sync version