Module anymap2::any

The different types of Any for use in a map.

This is based on std::any, but goes a little further, with CloneAny being a cloneable Any. CloneAnySend, CloneAnySync, and CloneAnySendSync further retrict what can be placed in the map with the Send and Sync bounds.

Traits

Any

See also CloneAny for a cloneable version of this trait.

CloneAny

See also Any for a version without the Clone requirement.

CloneAnySend

See also Any for a version without the Clone + Send requirements.

CloneAnySendSync

See also Any for a version without the Clone + Send + Sync requirements.

CloneAnySync

See also Any for a version without the Clone + Sync requirements.

UncheckedAnyExt